添加菜单审核接口 修改菜单删除逻辑

+++ src/main/java/com/cennavi/tp/service/impl/MenuDataServiceImpl.java
@@ -1,11 +1,17 @@
 package com.cennavi.tp.service.impl;
 
 import com.cennavi.tp.beans.MenuSubtitleBean;
+import com.cennavi.tp.beans.UserinfoBean;
+import com.cennavi.tp.common.result.Result;
+import com.cennavi.tp.common.result.ResultModel;
+import com.cennavi.tp.dao.ContentDao;
 import com.cennavi.tp.dao.MenuDataDao;
+import com.cennavi.tp.service.ContentService;
 import com.cennavi.tp.service.MenuDataService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
@@ -16,6 +22,8 @@ public class MenuDataServiceImpl implements MenuDataService {
 
     @Autowired
     private MenuDataDao menuDataDao;
+    @Autowired
+    private ContentService contentService;
 
     @Override
     public List<MenuSubtitleBean> getMenuList(Integer uid,String model) {
@@ -77,8 +85,31 @@ public class MenuDataServiceImpl implements MenuDataService {
     }
 
     @Override
-    public Integer deleteMenuSubtitleBeanById(Integer id) {
-        return menuDataDao.deleteMenuSubtitleBeanById(id);
+    public ResultModel deleteMenuSubtitleBeanById(Integer id, HttpServletRequest request) {
+        try{
+            //获取当前登录账号
+            UserinfoBean user = (UserinfoBean) request.getSession().getAttribute("user");
+            if(user==null) Result.buildUnLogin();
+            //获取菜单
+            MenuSubtitleBean menu = menuDataDao.getMenuSubtitleBeanById(id);
+            //菜单属于当前账号或者超级管理员才可以删除
+            if((menu.getUid()==user.getId() && menu.getStatus()!=1 && menu.getStatus()!=2) || user.getRole()==0){
+                //删除内容
+                contentService.deleteItemById(id);
+                //删除菜单
+                Integer index = menuDataDao.deleteMenuSubtitleBeanById(id);
+                if(index == 0){
+                    return Result.fail("输入的id无对应数据",index);
+                }else{
+                    return Result.success("删除成功",index);
+                }
+            }else{
+                return Result.fail("无删除权限");
+            }
+        }catch (Exception e){
+            e.getStackTrace();
+            return Result.build500("出现异常");
+        }
     }
 
     @Override
